RubiksCube converted to a single dialog app
In this version RubiksCube.cpp has been changed to be called by MainDlg and feed output back to it.

There is one fix in RubiksCube.cpp TreeNode::buildChildrenForBottomMap(). It handles the case where there is a collision when putting the node in the map, now it favors the one with the fewer number of moves.