Getting CppRestHttpGet to build in x64
As part of this I researched removing the include/lib paths for vcpkg packages. Vcpkg supports this, but it is not 100% flexible. You run 'vcpkg integrate install' from where vcpkg is built and it configures VS to use the include/lib folders for all VS projects on a per-Windows-user basis.

See: https://docs.microsoft.com/en-us/cpp/vcpkg